rename: popover, button_floating 컴포넌트명.tsx으로 파일명 수정
